// src/modules/deadlines/types/deadline-source.type.ts
export enum NotificationSourceType {
  PROJECT = 'PROJECT',
  PHASE = 'PHASE',
  TASK = 'TASK',
  MILESTONE = 'MILESTONE',
}

export enum NotificationType {
  MILESTONE_READY_FOR_VALIDATION = 'MILESTONE_READY_FOR_VALIDATION',

  MILESTONE_DUE_TOMORROW = 'MILESTONE_DUE_TOMORROW',
  TASK_DUE_TOMORROW = 'TASK_DUE_TOMORROW',
  PHASE_DUE_TOMORROW = 'PHASE_DUE_TOMORROW',
  PROJECT_DUE_TOMORROW = 'PROJECT_DUE_TOMORROW',
}